Thin down er-nix transitive dependencies.
Trying to move in the direction of relying more on the upstream caches.
We've diverged too much and consistently hit cachix cache misses, when
anything changes - we look at the long CI rebuild times (again, the
results of which get quickly evicted from the cachix and then people
see same on the laptops).

The intent is to integrate with haskell.nix (or even just nixpkgs)
more directly in the projects themselves and piggy back on publicly
cached binaries.

This project effectively remains a (hopefully tiny) collection of
Earnest-specific nix derivations.

* Flip nixpkgs to nixpkgs-unstable branch, we want to use versions
  that are likely to be in the upstream caches, not master.

* Drop okta-aws-login, it's distributed differnetly now.
  See updated https://github.com/EarnestResearch/eng-docs/tree/master/Okta#configure-aws-command-line-tools

* Drop haskell.nix and friends

* Drop custom skopeo hack, upstream has a much newer version, we should
  try that

* Drop custom niv version, upstream already matches what we want.

* Drop HLS stuff, again, let's rely on upstream caches